requirements.txt
vaderSentiment
httpx
fastapi
uvicorn
python-dotenv